Add files via upload
This program, asks the user for the input and output file names. It assumes that the input file contains the New York City Neighborhood Tabulation Areas (such as nynta.csv). I used SQL to filter the dataset to return the NTACode and NTAName columns from the input CSV file. I also saved the result as a CSV in the output file named by the user.